[Bronze I] Title: 유진수, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/1356. 유진수/유진수.cc
@@ -3,9 +3,9 @@
 #include <string>
 
 using namespace std;
-				 
+
 int main() {
-	unsigned long int N;
+    int N;
 	cin >> N;
 	string str = to_string(N);
